chore/refactor-shared-utils: refactor: reorganize utility functions and update imports for better structure and maintainability
