+++ src/shared/utils/index.ts
@@ -1,11 +1,9 @@
 export * from './account';
 export * from './code';
 export * from './debounce';
+export * from './fileUtils';
 export * from './git';
 export * from './oclif';
 export * from './prompts';
-export * from './runCommand';
-export * from './runMultipleCommand';
-export * from './runOnlyCommand';
+export * from './runner';
 export * from './versionFile';
-export * from './fileUtils';
