+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/cli-help.ts","./src/index.ts","./src/api/index.ts","./src/api/constants/colors.ts","./src/api/constants/index.ts","./src/api/directory/currentdirectory.ts","./src/api/directory/index.ts","./src/api/manifest/index.ts","./src/api/manifest/manifestinformation.ts","./src/api/modules/index.ts","./src/api/modules/prompts.ts","./src/api/table/index.ts","./src/api/table/render-table.ts","./src/api/table/table.ts","./src/commands/bash/run.ts","./src/commands/git/clone.ts","./src/commands/git/init.ts","./src/commands/git/release.ts","./src/commands/git/setup/conventional.ts","./src/commands/git/setup/devops/template.ts","./src/commands/pull-request/labels/suggest.ts","./src/commands/vtex/login.ts","./src/commands/vtex/run-script.ts","./src/commands/vtex/run.ts","./src/commands/vtex/cms/backup.ts","./src/commands/vtex/cms/deploy.ts","./src/hooks/init.ts","./src/modules/index.ts","./src/modules/utils.ts","./src/modules/apps/index.ts","./src/modules/apps/bash/index.ts","./src/modules/apps/bash/run/index.ts","./src/modules/apps/bash/run/utils.ts","./src/modules/apps/git/index.ts","./src/modules/apps/git/clone/index.ts","./src/modules/apps/git/clone/utils.ts","./src/modules/apps/git/init/index.ts","./src/modules/apps/git/init/utils.ts","./src/modules/apps/git/release/changelog.ts","./src/modules/apps/git/release/constants.ts","./src/modules/apps/git/release/createrelease.ts","./src/modules/apps/git/release/index.ts","./src/modules/apps/git/release/utils.ts","./src/modules/apps/git/setup/index.ts","./src/modules/apps/git/setup/conventional/index.ts","./src/modules/apps/git/setup/conventional/utils.ts","./src/modules/apps/git/setup/devops/index.ts","./src/modules/apps/git/setup/devops/templates/index.ts","./src/modules/apps/git/setup/devops/templates/utils.ts","./src/modules/apps/pull-request/index.ts","./src/modules/apps/pull-request/labels/index.ts","./src/modules/apps/pull-request/labels/suggest/analyzers.ts","./src/modules/apps/pull-request/labels/suggest/colors.ts","./src/modules/apps/pull-request/labels/suggest/formatters.ts","./src/modules/apps/pull-request/labels/suggest/index.ts","./src/modules/apps/pull-request/labels/suggest/types.ts","./src/modules/apps/pull-request/labels/suggest/utils.ts","./src/modules/apps/vtex/index.ts","./src/modules/apps/vtex/cms/backup.ts","./src/modules/apps/vtex/cms/index.ts","./src/modules/apps/vtex/cms/deploy/index.ts","./src/modules/apps/vtex/cms/deploy/utils.ts","./src/modules/apps/vtex/login/index.ts","./src/modules/apps/vtex/login/util/index.ts","./src/modules/apps/vtex/login/util/save-credentials.ts","./src/modules/apps/vtex/login/util/getauth/index.ts","./src/modules/apps/vtex/run/index.ts","./src/modules/apps/vtex/run/utils.ts","./src/modules/apps/vtex/run-script/index.ts","./src/shared/index.ts","./src/shared/logger.ts","./src/shared/verbose.ts","./src/shared/constants/commands.ts","./src/shared/constants/endpoints.ts","./src/shared/constants/index.ts","./src/shared/constants/paths.ts","./src/shared/models/index.ts","./src/shared/utils/account.ts","./src/shared/utils/debounce.ts","./src/shared/utils/git.ts","./src/shared/utils/index.ts","./src/shared/utils/oclif.ts","./src/shared/utils/prompts.ts","./src/shared/utils/runcommand.ts","./src/shared/utils/runmultiplecommand.ts","./src/shared/utils/runonlycommand.ts","./src/shared/utils/versionfile.ts","./src/shared/utils/code/index.ts","./src/shared/utils/code/git/index.ts","./src/shared/utils/code/git/init.ts","./src/shared/utils/code/git/setup/index.ts","./src/shared/utils/code/git/setup/conventional/index.ts","./src/shared/utils/code/git/setup/devops/index.ts","./src/shared/utils/code/git/setup/devops/templates/base.ts","./src/shared/utils/code/git/setup/devops/templates/index.ts","./src/shared/utils/fileutils/directory.ts","./src/shared/utils/fileutils/files.ts","./src/shared/utils/fileutils/index.ts","./src/shared/utils/fileutils/manifests.ts"],"version":"5.9.2"}
+{"root":["./src/cli-help.ts","./src/index.ts","./src/api/index.ts","./src/api/constants/colors.ts","./src/api/constants/index.ts","./src/api/directory/currentdirectory.ts","./src/api/directory/index.ts","./src/api/manifest/index.ts","./src/api/manifest/manifestinformation.ts","./src/api/modules/index.ts","./src/api/modules/prompts.ts","./src/api/table/index.ts","./src/api/table/render-table.ts","./src/api/table/table.ts","./src/commands/bash/run.ts","./src/commands/git/clone.ts","./src/commands/git/init.ts","./src/commands/git/release.ts","./src/commands/git/setup/conventional.ts","./src/commands/git/setup/devops/template.ts","./src/commands/pull-request/labels/suggest.ts","./src/commands/vtex/login.ts","./src/commands/vtex/run-script.ts","./src/commands/vtex/run.ts","./src/commands/vtex/cms/backup.ts","./src/commands/vtex/cms/deploy.ts","./src/hooks/init.ts","./src/modules/index.ts","./src/modules/utils.ts","./src/modules/apps/index.ts","./src/modules/apps/bash/index.ts","./src/modules/apps/bash/run/index.ts","./src/modules/apps/bash/run/utils.ts","./src/modules/apps/git/index.ts","./src/modules/apps/git/clone/index.ts","./src/modules/apps/git/clone/utils.ts","./src/modules/apps/git/init/index.ts","./src/modules/apps/git/init/utils.ts","./src/modules/apps/git/release/changelog.ts","./src/modules/apps/git/release/constants.ts","./src/modules/apps/git/release/createrelease.ts","./src/modules/apps/git/release/index.ts","./src/modules/apps/git/release/utils.ts","./src/modules/apps/git/setup/index.ts","./src/modules/apps/git/setup/conventional/index.ts","./src/modules/apps/git/setup/conventional/utils.ts","./src/modules/apps/git/setup/devops/index.ts","./src/modules/apps/git/setup/devops/templates/index.ts","./src/modules/apps/git/setup/devops/templates/utils.ts","./src/modules/apps/pull-request/index.ts","./src/modules/apps/pull-request/labels/index.ts","./src/modules/apps/pull-request/labels/suggest/analyzers.ts","./src/modules/apps/pull-request/labels/suggest/colors.ts","./src/modules/apps/pull-request/labels/suggest/formatters.ts","./src/modules/apps/pull-request/labels/suggest/index.ts","./src/modules/apps/pull-request/labels/suggest/types.ts","./src/modules/apps/pull-request/labels/suggest/utils.ts","./src/modules/apps/vtex/index.ts","./src/modules/apps/vtex/cms/backup.ts","./src/modules/apps/vtex/cms/index.ts","./src/modules/apps/vtex/cms/deploy/index.ts","./src/modules/apps/vtex/cms/deploy/utils.ts","./src/modules/apps/vtex/login/index.ts","./src/modules/apps/vtex/login/util/index.ts","./src/modules/apps/vtex/login/util/save-credentials.ts","./src/modules/apps/vtex/login/util/getauth/index.ts","./src/modules/apps/vtex/run/index.ts","./src/modules/apps/vtex/run/utils.ts","./src/modules/apps/vtex/run-script/index.ts","./src/shared/index.ts","./src/shared/logger.ts","./src/shared/verbose.ts","./src/shared/constants/commands.ts","./src/shared/constants/endpoints.ts","./src/shared/constants/index.ts","./src/shared/constants/paths.ts","./src/shared/models/index.ts","./src/shared/utils/account.ts","./src/shared/utils/debounce.ts","./src/shared/utils/git.ts","./src/shared/utils/index.ts","./src/shared/utils/oclif.ts","./src/shared/utils/prompts.ts","./src/shared/utils/runcommand.ts","./src/shared/utils/versionfile.ts","./src/shared/utils/code/index.ts","./src/shared/utils/code/git/index.ts","./src/shared/utils/code/git/init.ts","./src/shared/utils/code/git/setup/index.ts","./src/shared/utils/code/git/setup/conventional/index.ts","./src/shared/utils/code/git/setup/devops/index.ts","./src/shared/utils/code/git/setup/devops/templates/base.ts","./src/shared/utils/code/git/setup/devops/templates/index.ts","./src/shared/utils/fileutils/directory.ts","./src/shared/utils/fileutils/files.ts","./src/shared/utils/fileutils/index.ts","./src/shared/utils/fileutils/manifests.ts","./src/shared/utils/runner/index.ts","./src/shared/utils/runner/runcommand.ts","./src/shared/utils/runner/runmultiplecommand.ts","./src/shared/utils/runner/runonlycommand.ts"],"version":"5.9.2"}
